Refactor import paths from mathUtils to math and rename test file accordingly
