+++ src/bedrock/utils/math.ts
